+++ src/main/java/com/projeto/lista_tarefas/services/UserService.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -15,19 +14,24 @@ public class UserService {
     private UserRepository userRepository;
 
     public List<User> getAllUsers() {
-        return userRepository.findAll();
+            return userRepository.findAll();
     }
 
     public User getUserById(Long id){
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id).orElseThrow(()->new RuntimeException("Usuário não encontrado"));
     }
 
     public User createUser(User user) {
+
         return userRepository.save(user);
     }
 
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new RuntimeException("Usuário não encontrado");
+        }else{
+            userRepository.deleteById(id);
+        }
     }
